fix: move client methods out of CorruptedVaultHelper to CorruptedVaultClientHelper
